Refactor site into feature modules
