+++ src/features/exam-dashboard/components/index.ts
@@ -1,0 +1,13 @@
+export { default as Announcements } from "./Announcements";
+export { default as ConvocationGenerator } from "./ConvocationGenerator";
+export { default as ExamDashboard } from "./ExamDashboard";
+export { default as Header } from "./Header";
+export { default as Hero } from "./Hero";
+export { default as RoomSetup } from "./RoomSetup";
+export { default as RoomsStatus } from "./RoomsStatus";
+export { default as SurveillanceTable } from "./SurveillanceTable";
+export { default as DashboardFooter } from "./Footer";
+export { default as ExamDashboardPageLayout } from "./layout/ExamDashboardPageLayout";
+export { default as TodayBadge } from "./TodayBadge";
+export { default as TypeBadge } from "./TypeBadge";
+export * from "./NavigationButtons";
