+++ src/features/exam-dashboard/data/index.ts
@@ -1,0 +1,1 @@
+export * from "./dashboard-data";
