+++ src/features/exam-dashboard/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useDashboardTabs } from "./useDashboardTabs";
