+++ src/features/exam-dashboard/index.ts
@@ -1,0 +1,7 @@
+export { default as ExamDashboardPage } from "./pages/ExamDashboardPage";
+export * from "./components";
+export * from "./context";
+export * from "./data";
+export * from "./hooks";
+export * from "./services";
+export * from "./utils";
