+++ src/features/exam-dashboard/pages/ExamDashboardPage.tsx
@@ -1,0 +1,28 @@
+import {
+  Announcements,
+  BackToHomeButton,
+  ConvocationGenerator,
+  ExamDashboard,
+  ExamDashboardPageLayout,
+  Header,
+  Hero,
+  RoomSetup,
+  RoomsStatus,
+  SurveillanceTable,
+} from "../components";
+
+export default function ExamDashboardPage() {
+  return (
+    <ExamDashboardPageLayout action={<BackToHomeButton />}>
+      <Header />
+      <Hero />
+      <ExamDashboard>
+        <RoomSetup />
+        <SurveillanceTable />
+        <ConvocationGenerator />
+        <RoomsStatus />
+        <Announcements />
+      </ExamDashboard>
+    </ExamDashboardPageLayout>
+  );
+}
