+++ src/features/home/constants.ts
@@ -1,0 +1,18 @@
+export const HOME_PAGE_CONTENT = {
+  subtitle: "Espace examens blancs LFJP",
+  title:
+    "Toute l'organisation des examens blancs centralisée pour les enseignants du LFJP",
+  description:
+    "Cet espace réunit l'ensemble des informations pratiques nécessaires pour préparer, coordonner et faire vivre les épreuves. Vous y trouverez les documents, convocations, affectations de salles et les consignes indispensables pour guider sereinement chaque étape du déroulement des examens blancs.",
+};
+
+export const HOME_DASHBOARD_ENTRY = {
+  to: "/examens-blancs",
+  iconLabel: "Accéder à l'organisation des examens blancs",
+  subtitle: "Dessin interactif",
+  title: "Baccalauréat blanc 1ère et Terminale",
+  dateLabel: "10, 11 et 12 décembre 2025",
+  description:
+    "Cliquez pour accéder à la préparation détaillée des épreuves, aux répartitions des salles et aux outils pédagogiques.",
+  footerLabel: "Découvrir l'organisation",
+};
