--- src/lib/cn.ts
@@ -1,5 +1,0 @@
-export function cn(
-  ...classes: Array<string | false | null | undefined>
-): string {
-  return classes.filter(Boolean).join(" ");
-}
