+++ src/shared/components/Table.tsx
@@ -1,0 +1,32 @@
+import type {
+  HTMLAttributes,
+  TableHTMLAttributes,
+  TdHTMLAttributes,
+  ThHTMLAttributes,
+} from "react";
+
+import { cn } from "../lib";
+
+export function Table({ className, ...props }: TableHTMLAttributes<HTMLTableElement>) {
+  return <table className={cn("w-full text-left text-sm text-slate-600", className)} {...props} />;
+}
+
+export function TableHead({ className, ...props }: HTMLAttributes<HTMLTableSectionElement>) {
+  return <thead className={cn("bg-slate-100 text-xs uppercase text-slate-700", className)} {...props} />;
+}
+
+export function TableBody({ className, ...props }: HTMLAttributes<HTMLTableSectionElement>) {
+  return <tbody className={cn("align-top divide-y divide-slate-200", className)} {...props} />;
+}
+
+export function TableRow({ className, ...props }: HTMLAttributes<HTMLTableRowElement>) {
+  return <tr className={cn("transition-colors hover:bg-slate-50", className)} {...props} />;
+}
+
+export function TableHeaderCell({ className, ...props }: ThHTMLAttributes<HTMLTableCellElement>) {
+  return <th className={cn("px-6 py-3", className)} {...props} />;
+}
+
+export function TableCell({ className, ...props }: TdHTMLAttributes<HTMLTableCellElement>) {
+  return <td className={cn("px-6 py-4", className)} {...props} />;
+}
