+++ src/shared/lib/cn.ts
@@ -1,0 +1,5 @@
+export const cn = (
+  ...classes: Array<string | false | null | undefined>
+): string => classes.filter(Boolean).join(" ");
+
+export default cn;
